+++ database/plugin/blob/blob.go
@@ -15,7 +15,5 @@
 package blob
 
 import (
-	_ "github.com/blinklabs-io/dingo/database/plugin/blob/aws"
 	_ "github.com/blinklabs-io/dingo/database/plugin/blob/badger"
-	_ "github.com/blinklabs-io/dingo/database/plugin/blob/gcs"
 )
