+++ database/plugin/blob/gcs/commit_timestamp.go
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+//go:build dingo_extra_plugins
+
 package gcs
 
 import (
